Test (5)

+++ pymoo/problems/many/wfg.py
@@ -13,11 +13,11 @@ class WFG(Problem):
                          n_obj=n_obj,
                          n_constr=0,
                          xl=0.0,
-                         xu=2 * np.arange(1, n_var + 1).astype(np.float),
-                         type_var=np.double,
+                         xu=2 * np.arange(1, n_var + 1).astype(float),
+                         type_var=float,
                          **kwargs)
 
-        self.S = np.arange(2, 2 * self.n_obj + 1, 2).astype(np.float)
+        self.S = np.arange(2, 2 * self.n_obj + 1, 2).astype(float)
         self.A = np.ones(self.n_obj - 1)
 
         if k:
